fix(apu): Properly clear all audio registers when disabling APU
When writing to NR52 and disabled the APU, we should properly clear out
the sound channels' register. This includes not only the NRxy fields,
but also the other internal registers.

Having a dedicated function for clearing registers allows us to not
call `write_register` and unnecessarily pass the frame_step into the
function. We don't need to worry about any edge cases when writing 0,
because 0 doesn't cause any extra clocking since there cannot be a length
reload nor can there be a rising edge in the length bit.