refactor(CPU): Use `complete_cycle` instead of `fetch_cycle` in `step` fn
We don't need to worry about interrupt handling, since it's the first instruction
and it's impossible to have an interrupt here.

In fact, we really don't even need to use `complete_cycle` because we only need
to fetch the byte of the first instruction. We don't need to set the
`last_instruction` or reset the `cycle`.