fix(Timer): Implement TIMA ticking when writing to DIV and TAC
This allows us to pass `rapid_toggle`, the last failing test in the
`timer` folder for Mooneye.

However, this causes the `halt_ime1_timing2-GS` and `di_timing-GS` tests
to fail because we moved the interrupt check from T4 of the current instruction
to T3 of the next instruction (opcode fetch). Because of this, a VBlank
interrupt, that shouldn't be processed, ends up being processed.

I believe that this is still the right fix, in terms of accuracy, and
instead we should look into fixing it from the PPU side.